requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
keras==2.7.0
Keras-Preprocessing==1.1.2
gunicorn==20.1.0
librosa==0.9.2
matplotlib==3.5.2
tensorflow-cpu==2.7.0
opencv-python
imagen-pytorch==1.15.1